Refactoring. Add query class to module
